refactor(renderer): move _new_design UI primitives to ui/
First slice of dissolving the era-named _new_design/ folder. The seven
domain-free widgets — moss-dialog, moss-mini-button, moss-select-avatar(+
fancy), copy-hash, icons, defaultIcons — move to a new top-level ui/.

Pure file moves + import-path updates; no behaviour change. See
plans/elements-reorganization.md.
