+++ src/renderer/src/elements/main-dashboard.ts
@@ -48,7 +48,7 @@ import '../layout/views/asset-view.js';
 import '../groups/elements/group-container.js';
 import './debugging-panel/debugging-panel.js';
 
-import './_new_design/moss-dialog.js';
+import '../ui/moss-dialog.js';
 import './_new_design/moss-settings/moss-settings.js';
 import './design-feedback/design-feedback-controller.js';
 import './tool-info-dialog.js';
@@ -84,8 +84,8 @@ import {
   magnifyingGlassIcon,
   turingBlobIcon,
   turingBlobIconHover,
-} from './_new_design/icons.js';
-import { MossDialog } from './_new_design/moss-dialog.js';
+} from '../ui/icons.js';
+import { MossDialog } from '../ui/moss-dialog.js';
 
 type OpenTab =
   | {
