+++ src/renderer/src/moss-app.ts
@@ -17,10 +17,10 @@ import { MossStore } from './moss-store.js';
 import { appletDevConfig, getConductorInfo, ImportGroupsProgress } from './electron-api.js';
 import { LegacyProfileInfo } from './electron-api.js';
 import {localized, msg, str} from '@lit/localize';
-import { arrowLeftShortIcon, createGroupIcon, mossIcon } from './elements/_new_design/icons.js';
-import './elements/_new_design/moss-select-avatar.js';
-import './elements/_new_design/moss-select-avatar-fancy.js';
-import { defaultIcons } from './elements/_new_design/defaultIcons.js';
+import { arrowLeftShortIcon, createGroupIcon, mossIcon } from './ui/icons.js';
+import './ui/moss-select-avatar.js';
+import './ui/moss-select-avatar-fancy.js';
+import { defaultIcons } from './ui/defaultIcons.js';
 // import { GroupProfile } from '@theweave/api';
 import SlInput from '@shoelace-style/shoelace/dist/components/input/input.js';
 import { partialModifiersFromInviteLink } from '@theweave/utils';
