+++ src/renderer/src/personal-views/tool-library/elements/install-tool-dialog-web2.ts
@@ -23,8 +23,8 @@ import { GroupStore } from '../../../groups/group-store.js';
 import { mossStoreContext } from '../../../context.js';
 import { MossStore } from '../../../moss-store.js';
 import { ToolAndCurationInfo } from '../../../types.js';
-import { MossDialog } from '../../../elements/_new_design/moss-dialog.js';
-import '../../../elements/_new_design/moss-dialog.js';
+import { MossDialog } from '../../../ui/moss-dialog.js';
+import '../../../ui/moss-dialog.js';
 import {Applet, AppletAgent} from "@theweave/group-client";
 import { toolCompatibilityIdFromDistInfoString } from '@theweave/utils';
 import { DistributionInfo, TDistributionInfo, ToolInfoAndVersions } from '@theweave/moss-types';
