+++ src/renderer/src/ui/copy-hash.ts
