+++ src/renderer/src/ui/defaultIcons.ts
