+++ src/renderer/src/ui/icons.ts
